[update] Add Model Edit Container
