+++ src/webapp/styles/common/_utils.scss
@@ -2,6 +2,10 @@
   display: block;
 }
 
+.cb-relative {
+  position: relative;
+}
+
 .cb-text-center {
   text-align: center;
 }
